requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
requests==2.32.0
APScheduler==3.10.1
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1